# wls-gui-admintool/src/resources/openapis/openapi.basisdaten.0.4.0.json
{"openapi":"3.1.0","info":{"title":"Basisdatenservice","contact":{"name":"Your Name","email":"Your E-Mail-Address"},"version":"0.4.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/businessActions/wahltermindaten/{wahltagID}":{"put":{"tags":["wahltermindaten-controller"],"description":"Triggert die Einrichtung der Wahltermindaten.","operationId":"putWahltermindaten","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Die Wahltermindaten werden angelegt."},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}},"delete":{"tags":["wahltermindaten-controller"],"description":"Löscht die Wahltermindaten eines bestimmten Wahltags.","operationId":"deleteWahltermindaten","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Die Wahltermindaten werden gelöscht."},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/wahlen/{wahltagID}":{"get":{"tags":["wahlen-controller"],"description":"Laden der Wahlen des Wahltages { wahltagID }.","operationId":"getWahlen","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Wahlen des Wahltags erfolgreich zurückgegeben","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WahlDTO"}}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}},"post":{"tags":["wahlen-controller"],"description":"Speichern einer Liste von Wahlen.","operationId":"postWahlen","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WahlDTO"}}}},"required":true},"responses":{"200":{"description":"Wahlen erfolgreich abgespeichert."},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/ungueltigews/{wahltagID}/{wahlbezirksart}":{"get":{"tags":["ungueltige-wahlscheine-controller"],"description":"Abrufen der ungueltigen Wahlscheine eines Wahltages für eine bestimmte Wahlbezirksart. Kommen als Anhang im csv-Format mit den Spalten Nachname, Vorname und Nummer","operationId":"getUngueltigeWahlscheine","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirksart","in":"path","required":true,"schema":{"type":"string","enum":["UWB","BWB"]}}],"responses":{"200":{"description":"Ungueltige Wahlscheine erfolgreich zurückgegeben.","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}},"post":{"tags":["ungueltige-wahlscheine-controller"],"description":"Speichern der ungueltigen Wahlscheine eines Wahltages für eine bestimmte Wahlbezirksart","operationId":"setUngueltigeWahlscheine","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirksart","in":"path","required":true,"schema":{"type":"string","enum":["UWB","BWB"]}}],"responses":{"200":{"description":"Ungueltige Wahlscheine erfolgreich gespeichert."},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/resetWahlen":{"post":{"tags":["wahlen-controller"],"description":"Setzt die Attribute Farbe, Reihenfolge und Waehlerverzeichnis der vorhandenen Wahlen auf die Standardwerte.","operationId":"resetWahlen","responses":{"200":{"description":"Die Wahlen wurden zurückgesetzt."},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/handbuch/{wahltagID}/{wahlbezirksart}":{"get":{"tags":["handbuch-controller"],"description":"Abrufen des Handbuches einer Wahl für eine bestimmte Wahlbezirksart","operationId":"getHandbuch","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirksart","in":"path","required":true,"schema":{"type":"string","enum":["UWB","BWB"]}}],"responses":{"200":{"description":"Handbuch erfolgreich zurückgegeben.","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}},"post":{"tags":["handbuch-controller"],"description":"Speichern eines Handbuches einer Wahl für eine bestimmte Wahlbezirksart","operationId":"setHandbuch","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirksart","in":"path","required":true,"schema":{"type":"string","enum":["UWB","BWB"]}}],"responses":{"200":{"description":"Handbuch erfolgreich gespeichert"},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/wahlvorschlaege/{wahlID}/{wahlbezirkID}":{"get":{"tags":["wahlvorschlaege-controller"],"description":"Laden der Wahlvorschlaege des Wahllokals {wahlbezirkID} für die Wahl {wahlID}.","operationId":"getWahlvorschlaege","parameters":[{"name":"wahlID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirkID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Wahlvorschläge erfolgreich zurückgegeben.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WahlvorschlaegeDTO"}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/wahltage":{"get":{"tags":["wahltage-controller"],"description":"Laden der Liste der Wahltage, aufsteigend sortiert nach Wahltag-Datum und nicht länger als 3 Monate in der Vergangenheit.","operationId":"getWahltage","responses":{"200":{"description":"Wahltage erfolgreich zurückgegeben.","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WahltagDTO"}}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/wahlbezirke/{wahltagID}":{"get":{"tags":["wahlbezirke-controller"],"description":"Laden der Liste der Wahlbezirke, die einem vorgegebenen Wahltag (Parameter wahltagID) entsprechen.","operationId":"getWahlbezirke","parameters":[{"name":"wahltagID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Wahlbezirke erfolgreich zurückgegeben.","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WahlbezirkDTO"}}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/referendumvorlagen/{wahlID}/{wahlbezirkID}":{"get":{"tags":["referendumvorlagen-controller"],"description":"Laden der Referendumsvorlagen des Wahllokals {wahlbezirkID} für eine Wahl {wahlID}.","operationId":"getReferendumvorlagen","parameters":[{"name":"wahlID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirkID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Referendumvorlage erfolgreich zurückgegeben","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReferendumvorlagenDTO"}}}},"204":{"description":"Keine Daten vom Fremdsystem geliefert"},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/kopfdaten/{wahlID}/{wahlbezirkID}":{"get":{"tags":["kopfdaten-controller"],"description":"Laden der Kopfdaten für die Wahl {wahlID} für den Wahlbezirk {wahlbezirkID}.","operationId":"getKopfdaten","parameters":[{"name":"wahlID","in":"path","required":true,"schema":{"type":"string"}},{"name":"wahlbezirkID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Kopfdaten erfolgreich zurückgegeben.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/KopfdatenDTO"}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}},"/businessActions/asyncProgress":{"get":{"tags":["async-progress-controller"],"description":"Abrufen des Fortschrittes bei der Importierung der Wahltermindaten.","operationId":"getAsyncProgress","responses":{"200":{"description":"Importfortschritt wurde geliefert.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AsyncProgressDTO"}}}},"400":{"description":"Validierung des Requestbody fehlgeschlagen","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}},"404":{"description":"Ressource nicht gefunden"},"500":{"description":"Unbehandelter interner Fehler, z.B. Kommunikation mit anderen Systemen oder andere Arten von unbehandelten Fehlern","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WlsExceptionDTO"}}}}}}}},"components":{"schemas":{"FarbeDTO":{"type":"object","properties":{"r":{"type":"integer","format":"int64","maximum":255,"minimum":0},"g":{"type":"integer","format":"int64","maximum":255,"minimum":0},"b":{"type":"integer","format":"int64","maximum":255,"minimum":0}},"required":["b","g","r"]},"WahlDTO":{"type":"object","properties":{"wahlID":{"type":"string"},"name":{"type":"string"},"reihenfolge":{"type":"integer","format":"int64"},"waehlerverzeichnisNummer":{"type":"integer","format":"int64"},"wahltag":{"type":"string","format":"date"},"wahlart":{"type":"string","enum":["BAW","BEB","BTW","BZW","EUW","LTW","MBW","OBW","SRW","SVW","VE"]},"farbe":{"$ref":"#/components/schemas/FarbeDTO"},"nummer":{"type":"string"}},"required":["farbe","name","reihenfolge","waehlerverzeichnisNummer","wahlID","wahlart","wahltag"]},"KandidatDTO":{"type":"object","properties":{"identifikator":{"type":"string"},"name":{"type":"string"},"listenposition":{"type":"integer","format":"int64"},"direktkandidat":{"type":"boolean"},"tabellenSpalteInNiederschrift":{"type":"integer","format":"int64"},"einzelbewerber":{"type":"boolean"}},"required":["direktkandidat","einzelbewerber","identifikator","listenposition","name","tabellenSpalteInNiederschrift"]},"WahlvorschlaegeDTO":{"type":"object","properties":{"wahlID":{"type":"string"},"wahlbezirkID":{"type":"string"},"stimmzettelgebietID":{"type":"string"},"wahlvorschlaege":{"type":"array","items":{"$ref":"#/components/schemas/WahlvorschlagDTO"},"uniqueItems":true}},"required":["stimmzettelgebietID","wahlID","wahlbezirkID","wahlvorschlaege"]},"WahlvorschlagDTO":{"type":"object","properties":{"identifikator":{"type":"string"},"ordnungszahl":{"type":"integer","format":"int64"},"kurzname":{"type":"string"},"erhaeltStimmen":{"type":"boolean"},"kandidaten":{"type":"array","items":{"$ref":"#/components/schemas/KandidatDTO"},"uniqueItems":true}},"required":["erhaeltStimmen","identifikator","kurzname","ordnungszahl"]},"WahltagDTO":{"type":"object","properties":{"wahltagID":{"type":"string"},"wahltag":{"type":"string","format":"date"},"beschreibung":{"type":"string"},"nummer":{"type":"string"}},"required":["wahltag","wahltagID"]},"WahlbezirkDTO":{"type":"object","properties":{"wahlbezirkID":{"type":"string"},"wahltag":{"type":"string","format":"date"},"nummer":{"type":"string"},"wahlbezirkart":{"type":"string","enum":["UWB","BWB"]},"wahlnummer":{"type":"string"},"wahlID":{"type":"string"}},"required":["nummer","wahlID","wahlbezirkID","wahlbezirkart","wahlnummer","wahltag"]},"ReferendumoptionDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"position":{"type":"integer","format":"int64"}},"required":["id","name"]},"ReferendumvorlageDTO":{"type":"object","properties":{"wahlvorschlagID":{"type":"string"},"ordnungszahl":{"type":"integer","format":"int64"},"kurzname":{"type":"string"},"frage":{"type":"string"},"referendumoptionen":{"type":"array","items":{"$ref":"#/components/schemas/ReferendumoptionDTO"},"uniqueItems":true}},"required":["frage","kurzname","ordnungszahl","referendumoptionen","wahlvorschlagID"]},"ReferendumvorlagenDTO":{"type":"object","properties":{"stimmzettelgebietID":{"type":"string"},"referendumvorlage":{"type":"array","items":{"$ref":"#/components/schemas/ReferendumvorlageDTO"},"uniqueItems":true}},"required":["referendumvorlage","stimmzettelgebietID"]},"KopfdatenDTO":{"type":"object","properties":{"wahlID":{"type":"string"},"wahlbezirkID":{"type":"string"},"gemeinde":{"type":"string"},"stimmzettelgebietsart":{"type":"string","enum":["SB","SG","SK","WK"]},"stimmzettelgebietsnummer":{"type":"string"},"stimmzettelgebietsname":{"type":"string"},"wahlname":{"type":"string"},"wahlbezirknummer":{"type":"string"}},"required":["gemeinde","stimmzettelgebietsart","stimmzettelgebietsname","stimmzettelgebietsnummer","wahlID","wahlbezirkID","wahlbezirknummer","wahlname"]},"AsyncProgressDTO":{"type":"object","properties":{"forWahltag":{"type":"string","format":"date"},"wahlNummer":{"type":"string"},"lastStartTime":{"type":"string","format":"date-time"},"lastFinishTime":{"type":"string","format":"date-time"},"wahlvorschlaegeLoadingActive":{"type":"boolean"},"wahlvorschlaegeTotal":{"type":"integer","format":"int32"},"wahlvorschlageFinished":{"type":"integer","format":"int32"},"wahlvorschlaegeNext":{"type":"string"},"referendumLoadingActive":{"type":"boolean"},"referendumVorlagenTotal":{"type":"integer","format":"int32"},"referendumVorlagenFinished":{"type":"integer","format":"int32"},"referendumVorlagenNext":{"type":"string"}}},"WlsExceptionDTO":{"properties":{"category":{"type":"string","enum":["F","T","S","I"]},"code":{"type":"string"},"service":{"type":"string"},"message":{"type":"string"}},"required":["category","code","message","service"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}